Added code to make the JSON better. Added CSS, experimenting with table output. Might try to do appending divs for each new element?
So, i wasn't formatting the JSON properly, I was but not in the best manner. I was using literal rather than object. So i researched and found a method that applied keys then dived up the messages better. Muc better way to create json. the original way wasnt bad but it made it harder to output. Upnext, style and staggering messages. Also added CSS file.