Added Keyup, Edited UI
Man, jquery makes everything a lot easier, i need to really learn it. This update was mostly visual, i added some more style and removed code which was commented out. A lot of it was from testing it. The UI looks cleaner and actually like a chat bot. more to come.